feat: move completion session/menu to dispatcher helpers, update shell + CLI to use them
Agent-Logs-Url: https://github.com/microsoft/TypeAgent/sessions/8766b4b5-3498-4828-8c69-9f701f721f2a
