+++ ts/packages/dispatcher/dispatcher/package.json
@@ -19,7 +19,7 @@
     "./helpers/config": "./dist/helpers/config.js",
     "./helpers/status": "./dist/helpers/status.js",
     "./helpers/command": "./dist/helpers/command.js",
-    "./helpers/completion": "./dist/helpers/completion.js",
+    "./helpers/completion": "./dist/helpers/completion/index.js",
     "./internal": "./dist/internal.js",
     "./explorer": "./dist/explorer.js"
   },
