+++ ts/packages/shell/src/preload/electronTypes.ts
@@ -32,18 +32,7 @@ export type ClientActions =
 
 // end duplicate type section
 
-export type SearchMenuPosition = {
-    left: number;
-    bottom: number;
-};
-
-export type SearchMenuItem = {
-    matchText: string;
-    emojiChar?: string | undefined;
-    sortIndex?: number;
-    selectedText: string;
-    needQuotes?: boolean | undefined; // default is true, and will add quote to the selectedText if it has spaces.
-};
+export type { SearchMenuPosition, SearchMenuItem } from "agent-dispatcher/helpers/completion";
 
 export type SearchMenuUIUpdateData = {
     position?: SearchMenuPosition;
